@@ -4,7 +4,7 @@ buildscript {
         mavenCentral()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:7.3.1'
+        classpath 'com.android.tools.build:gradle:8.1.3'
     }
 }
 
@@ -16,22 +16,21 @@ allprojects {
     }
 }
 
-
 apply plugin: 'com.android.application'
 
 dependencies {
     implementation 'org.greenrobot:eventbus:3.0.0'
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
-    implementation 'androidx.annotation:annotation:1.4.0'
-    implementation 'androidx.appcompat:appcompat:1.4.2'
-    implementation "androidx.lifecycle:lifecycle-viewmodel:2.5.0"
-    implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:2.5.0"
-    implementation 'androidx.preference:preference:1.2.0'
+    implementation 'androidx.annotation:annotation:1.7.0'
+    implementation 'androidx.appcompat:appcompat:1.6.1'
+    implementation "androidx.lifecycle:lifecycle-viewmodel:2.6.2"
+    implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:2.6.2"
+    implementation 'androidx.preference:preference:1.2.1'
     implementation 'com.google.code.gson:gson:2.8.6'
     implementation 'com.google.android.flexbox:flexbox:3.0.0'
-    implementation 'com.google.android.material:material:1.6.1'
+    implementation 'com.google.android.material:material:1.10.0'
     implementation 'com.android.billingclient:billing:6.0.1'
-    implementation 'androidx.fragment:fragment:1.5.0'
+    implementation 'androidx.fragment:fragment:1.6.2'
 
     // implementation 'uk.co.samuelwall:material-tap-target-prompt:3.0.0'
     // implementation project(':externals:ShowcaseView:library')
@@ -41,7 +40,7 @@ dependencies {
     implementation fileTree(include: '*.jar', dir: 'libs')
     implementation 'com.github.prolificinteractive:material-calendarview:1.6.1'
     implementation 'androidx.palette:palette:1.0.0'
-    implementation 'androidx.exifinterface:exifinterface:1.3.3'
+    implementation 'androidx.exifinterface:exifinterface:1.3.6'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
     implementation "androidx.coordinatorlayout:coordinatorlayout:1.2.0"
     implementation 'androidx.mediarouter:mediarouter:1.3.0'
@@ -52,9 +51,9 @@ dependencies {
     implementation 'com.google.android.exoplayer:exoplayer:2.18.0'
     implementation 'com.github.android:renderscript-intrinsics-replacement-toolkit:b6363490c3'
     implementation 'com.github.martin-stone:hsv-alpha-color-picker-android:3.0.1'
-    implementation 'androidx.work:work-runtime:2.7.1'
-    implementation 'androidx.work:work-runtime-ktx:2.7.1'
-    implementation 'androidx.work:work-multiprocess:2.7.1'
+    implementation 'androidx.work:work-runtime:2.8.1'
+    implementation 'androidx.work:work-runtime-ktx:2.8.1'
+    implementation 'androidx.work:work-multiprocess:2.8.1'
     implementation 'com.prof18.rssparser:rssparser:4.0.2'
 
 
@@ -76,7 +75,8 @@ def getApiKeyDarkSky() {
 }
 
 android {
-    compileSdkVersion 33
+    namespace 'com.firebirdberlin.nightdream'
+    compileSdkVersion 34
 
     compileOptions {
         sourceCompatibility 1.8
@@ -130,7 +130,7 @@ android {
         minSdkVersion 17
         buildConfigField "String", "API_KEY_OWM", "\""+getApiKeyOWM()+"\""
         buildConfigField "String", "API_KEY_DARK_SKY", "\""+getApiKeyDarkSky()+"\""
-        targetSdkVersion 33
+        targetSdkVersion 34
         renderscriptTargetApi 19
         renderscriptSupportModeEnabled true
         vectorDrawables.useSupportLibrary = true
@@ -138,5 +138,7 @@ android {
     }
     buildFeatures {
         dataBinding true
+        renderScript true
+        aidl true
     }
 }
